Split qr and datamatrix (#35)
* moves decoding from general decoder to dedicated qr decoder

* moves static methods used for decoding to qr_decoder package as functions

* makes qr detection a separate function and detectCodeType a controller

* moves QR decoding stuff to its subdirectory

* moves QR detection to a separate package

---------
