+++ internal/decoding/QR/common/bit_reader/bit_reader_test.go
