+++ internal/decoding/QR/common/data_formats/data_formats.go
