+++ internal/decoding/QR/qr_v1/sequence.go
