+++ internal/decoding/QR/qr_v3/qr_v3.go
@@ -1,7 +1,7 @@
 package qr_v3
 
 import (
-	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/decoding/qr_v1"
+	qrdecoder "github.com/ChainsAre2Tight/qr-decoder-utils/internal/decoding/QR/common/qr_decoder"
 	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/interfaces"
 	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/types"
 	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/utils"
@@ -28,16 +28,6 @@ func (QRVer3) OOB() interfaces.OutOfBoundsInterface {
 	return oob{}
 }
 
-// Inherited from QRVer1
-func (QRVer3) ReadMetadata(matrix [][]bool) (interfaces.ModeInterface, interfaces.MaskInterface, error) {
-	return qr_v1.QRVer1{}.ReadMetadata(matrix)
-}
-
-// Inherited from QRVer1
-func (QRVer3) ReadFormat(matrix [][]bool, mask interfaces.MaskInterface, reader interfaces.BitReaderInterface) (interfaces.FormatInterface, error) {
-	return qr_v1.QRVer1{}.ReadFormat(matrix, mask, reader)
-}
-
 type oob struct{}
 
 func (oob) SkipCell(x, y int) bool {
@@ -50,4 +40,8 @@ func (oob) SkipCell(x, y int) bool {
 
 func (oob) SkipColumn(x int) bool {
 	return x == 6
+}
+
+func (q QRVer3) Decode(matrix [][]bool) (string, error) {
+	return qrdecoder.DecodeQR(matrix, q)
 }
