+++ internal/output/output.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"reflect"
 
-	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/decoding/common/masks"
+	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/decoding/QR/common/masks"
 	"github.com/ChainsAre2Tight/qr-decoder-utils/internal/utils"
 	"github.com/tealeg/xlsx"
 )
